Integrasi Midtrans

+++ app/Http/Controllers/Web/CategoryController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Web;
 
+use App\Models\Product;
 use App\Models\Category;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
@@ -86,8 +87,13 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->delete();
-
-        return redirect()->route('categories.index')->with('success', 'Category deleted successfully.');
+        $products = Product::where('category_id', $category->id)->first();
+        if($products){
+            return back()->with(['error' => 'Category already used']);
+        }
+        else{
+            $category->delete();
+            return redirect()->route('categories.index')->with('success', 'Category deleted successfully.');
+        }
     }
 }
